test: refactor reset all stores
